feat(dashboard): polish meeting cards and empty states
- Replace custom chip spans with sidebar-style pill tags
- Move recording/transcripts/AI icons inline with time, separated by dots
- Abbreviate weekday in time display (Mon vs Monday)
- Add truncation-aware tooltip on meeting title
- Remove tooltip from pending-actions description
- Add pTooltip with recurrence summary to Recurring chip
- Fix fullSummary to handle empty description (filter+join)
- Redesign date badge as split calendar (month strip + day body)
- Darken neutral gradient header; use bg-blue-500 for accent month strip
- Fix visibility dot icon centering with leading-none block
- Change private badge icon to fa-solid fa-shield-halved
- Redesign empty states: icon-left layout, gradient backgrounds, no header
- Polish empty state copy and icons (box-check for pending actions)
- Remove isDevMode mock data to show empty states for preview

+++ apps/lfx-one/src/app/modules/dashboards/components/dashboard-meeting-card/dashboard-meeting-card.component.ts
@@ -181,7 +181,7 @@ export class DashboardMeetingCardComponent {
       try {
         const date = new Date(startTime);
         if (isNaN(date.getTime())) return startTime;
-        const weekday = date.toLocaleDateString('en-US', { weekday: 'long' });
+        const weekday = date.toLocaleDateString('en-US', { weekday: 'short' });
         const time = date.toLocaleTimeString('en-US', { hour: 'numeric', minute: '2-digit', hour12: true });
         const duration = this.meetingDuration();
         return duration > 0 ? `${weekday}, ${time} · ${duration}m` : `${weekday}, ${time}`;
